Fixed mesh loading crashes on Linux.
These bugs are really hard to spot with the stack trace indicating
issues *deep* inside Eigen.
The issue was presented on the screen all along, with a misleading
underline on a seemingly "uninvolved" variable:
src/scene.cpp:163:11: warning: control reaches end of non-void function [-Wreturn-type]
  ifstream sceneFileHandle, constraintFileHandle;
           ^~~~~~~~~~~~~~~
All in all, we need to learn to read (and some -Werror...).